feat: 대진표 생성 기능 구현

+++ badminton-api/src/main/java/org/badminton/api/league/model/dto/LeagueUpdateRequest.java
@@ -36,7 +36,7 @@ public record LeagueUpdateRequest(
 	LocalDateTime closedAt,
 
 	@Schema(description = "참가 인원", example = "16")
-	Long playerCount,
+	int playerCount,
 
 	@Schema(description = "매칭 조건", example = "TIER")
 	String matchingRequirement
